Second try of final task

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        int temperatureFahrenheit=temperatureCelsius+32;
+       double temperatureFahrenheit=temperatureCelsius+32.0;
         System.out.println(temperatureFahrenheit);
     }
 }
